+++ fdr-node/fdr-node-all/src/assembly/fdr-stop.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+# 程序的根目录
+basedir=/usr/local/fdr-maintain
+
+PID=$(cat $basedir/fdr.pid)
+kill "$PID"