+++ fdr-node/fdr-node-all/src/main/java/com/dwarfeng/fdr/node/all/handler/LauncherSettingHandler.java
@@ -1,0 +1,28 @@
+package com.dwarfeng.fdr.node.all.handler;
+
+import com.dwarfeng.subgrade.stack.handler.Handler;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+
+@Component
+public class LauncherSettingHandler implements Handler {
+
+    @Value("${launcher.reset_filter_support}")
+    private boolean resetFilterSupport;
+    @Value("${launcher.reset_trigger_support}")
+    private boolean resetTriggerSupport;
+    @Value("${launcher.start_record_delay}")
+    private long startRecordDelay;
+
+    public boolean isResetFilterSupport() {
+        return resetFilterSupport;
+    }
+
+    public boolean isResetTriggerSupport() {
+        return resetTriggerSupport;
+    }
+
+    public long getStartRecordDelay() {
+        return startRecordDelay;
+    }
+}