+++ fdr-node/fdr-node-all/src/main/java/com/dwarfeng/fdr/node/all/launcher/Launcher.java
@@ -1,7 +1,8 @@
-package com.dwarfeng.fdr.node.maintain.launcher;
+package com.dwarfeng.fdr.node.all.launcher;
 
-import com.dwarfeng.fdr.node.maintain.handler.LauncherSettingHandler;
+import com.dwarfeng.fdr.node.all.handler.LauncherSettingHandler;
 import com.dwarfeng.fdr.stack.service.FilterSupportMaintainService;
+import com.dwarfeng.fdr.stack.service.RecordQosService;
 import com.dwarfeng.fdr.stack.service.TriggerSupportMaintainService;
 import com.dwarfeng.springterminator.stack.handler.Terminator;
 import com.dwarfeng.subgrade.stack.exception.ServiceException;
@@ -13,7 +14,7 @@ import org.springframework.context.support.ClassPathXmlApplicationContext;
  * 程序启动器。
  *
  * @author DwArFeng
- * @since 0.0.1-alpha
+ * @since 1.4.2
  */
 public class Launcher {
 
@@ -23,6 +24,7 @@ public class Launcher {
         ClassPathXmlApplicationContext ctx = new ClassPathXmlApplicationContext("classpath:spring/application-context*.xml");
         ctx.registerShutdownHook();
         ctx.start();
+        //判断是否重置触发器和过滤器。
         LauncherSettingHandler launcherSettingHandler = ctx.getBean(LauncherSettingHandler.class);
         if (launcherSettingHandler.isResetFilterSupport()) {
             LOGGER.info("重置过滤器支持...");
@@ -40,6 +42,29 @@ public class Launcher {
                 maintainService.reset();
             } catch (ServiceException e) {
                 LOGGER.warn("过滤器支持重置失败，异常信息如下", e);
+            }
+        }
+        // 判断是否开启记录服务。
+        long startRecordDelay = launcherSettingHandler.getStartRecordDelay();
+        RecordQosService recordQosService = ctx.getBean(RecordQosService.class);
+        if (startRecordDelay == 0) {
+            LOGGER.info("立即启动记录服务...");
+            try {
+                recordQosService.startRecord();
+            } catch (ServiceException e) {
+                LOGGER.error("无法启动记录服务，异常原因如下", e);
+            }
+        } else if (startRecordDelay > 0) {
+            LOGGER.info(startRecordDelay + " 毫秒后启动记录服务...");
+            try {
+                Thread.sleep(startRecordDelay);
+            } catch (InterruptedException ignored) {
+            }
+            LOGGER.info("启动记录服务...");
+            try {
+                recordQosService.startRecord();
+            } catch (ServiceException e) {
+                LOGGER.error("无法启动记录服务，异常原因如下", e);
             }
         }
         Terminator terminator = ctx.getBean(Terminator.class);